# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Qmoon",
    "description": "### 基本约束\n  1. 服务正常情况下，http状态码为`200`\n  1. 未登录,http状态码为`401`\n  1. 未授权,http状态码为`403`\n  1. 响应体为json对象\n\n---\n## http响应状态码说明\n| http status | 描述 |\n| - | :-: |\n| 200 | 服务正常，具体业务是否正常根据响应体里的code来判断|\n| 401 | 未登录|\n| 403 | 未授权|\n| 500 | 服务错误|\n",
    "version": "0.2"
  },
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "APIKeyHeader": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "parameters": {
    "chainIDParam": {
      "in": "path",
      "name": "chainId",
      "required": true,
      "type": "string",
      "description": "chain id"
    },
    "minHeightParam": {
      "in": "query",
      "name": "minHeight",
      "required": false,
      "type": "integer",
      "description": "最小高度"
    },
    "maxHeightParam": {
      "in": "query",
      "name": "maxHeight",
      "required": false,
      "type": "integer",
      "description": "最大高度"
    },
    "offsetParam": {
      "in": "query",
      "name": "offset",
      "required": false,
      "type": "integer",
      "description": "偏移"
    },
    "limitParam": {
      "in": "query",
      "name": "limit",
      "required": false,
      "type": "integer",
      "description": "数量"
    }
  },
  "paths": {
...
    "/nodes/{chainId}/status": {
      "get": {
        "summary": "链状态",
        "tags": [
          "baseApi"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/chainIDParam"
          }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/BaseResponseBody"
                },
                {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "object",
                      "properties": {
                        "last_height": {
                          "type": "integer",
                          "description": "最新高度"
                        },
                        "blocks": {
                          "type": "array",
                          "items": {
                            "$ref": "#/definitions/Status"
                          }
                        }
                      }
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
    "/nodes/{chainId}/blockchain": {
      "get": {
        "summary": "块链",
        "tags": [
          "baseApi"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/chainIDParam"
          },
          {
            "$ref": "#/parameters/minHeightParam"
          },
          {
            "$ref": "#/parameters/maxHeightParam"
          },
          {
            "$ref": "#/parameters/offsetParam"
          },
          {
            "$ref": "#/parameters/limitParam"
          }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/BaseResponseBody"
                },
                {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "object",
                      "properties": {
                        "last_height": {
                          "type": "integer",
                          "description": "最新高度"
                        },
                        "blocks": {
                          "type": "array",
                          "items": {
                            "$ref": "#/definitions/Blockchain"
                          }
                        }
                      }
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
...
    "/nodes/{chainId}/tx": {
      "get": {
        "summary": "交易详情",
        "tags": [
          "baseApi"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/chainIDParam"
          },
          {
            "in": "query",
            "name": "height",
            "required": false,
            "type": "integer",
            "description": "块高度"
          },
          {
            "in": "query",
            "name": "index",
            "required": false,
            "type": "integer",
            "description": "交易在当前块序号"
          }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/BaseResponseBody"
                },
                {
                  "type": "object",
                  "properties": {
                    "result": {
                      "$ref": "#/definitions/Tx"
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
...
    "/nodes/{chainId}/validators": {
      "get": {
        "summary": "validators",
        "tags": [
          "baseApi"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/chainIDParam"
          },
          {
            "in": "query",
            "name": "height",
            "required": false,
            "type": "integer",
            "description": "块高度"
          }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/BaseResponseBody"
                },
                {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "array",
                      "items": {
                        "$ref": "#/definitions/Validator"
                      }
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "BaseResponseBody": {
      "type": "object",
      "properties": {
        "jsonrpc": {
          "type": "string",
          "description": "jsonrpc"
        },
        "id": {
          "type": "string",
          "description": "jsonrpc"
        }
      }
    },
...
    "Blockchain": {
      "type": "object",
      "description": "块链信息",
      "properties": {
        "chinId": {
          "type": "string",
          "description": "chain id"
        },
        "height": {
          "type": "integer",
          "description": "高度"
        },
        "num_txs": {
          "type": "string",
          "description": "当前块交易数量"
        },
        "total_txs": {
          "type": "string",
          "description": "累计交易数量"
        },
        "time": {
          "type": "string",
          "description": "chain id"
        },
        "data_hash": {
          "type": "string",
          "description": "数据hash"
        },
        "validators_hash": {
          "type": "string",
          "description": "validators hash"
        }
      }
    },
...
    "TxCosmosDelegate": {
      "type": "object",
      "description": "交易信息",
      "properties": {
        "delegator_address": {
          "type": "string",
          "description": "delegator_address"
        },
        "validator_address": {
          "type": "string",
          "description": "delegator_address"
        },
        "value": {
          "type": "object",
          "description": "value",
          "properties": {
            "denom": {
              "type": "string"
            },
            "amount": {
              "type": "string"
            }
          }
        }
      }
    },
    "TxCosmosSend": {
      "type": "object",
      "description": "交易信息",
      "properties": {
        "from_address": {
          "type": "string"
        },
        "to_address": {
          "type": "string"
        },
        "amount": {
          "type": "object",
          "properties": {
            "denom": {
              "type": "string"
            },
            "amount": {
              "type": "string"
            }
          }
        }
      }
    },
    "TxCosmosWithdrawDelegatorReward": {
      "type": "object",
      "description": "交易信息",
      "properties": {
        "delegator_address": {
          "type": "string"
        },
        "validator_address": {
          "type": "string"
        }
      }
    },
    "TxCosmosBeginUnbonding": {
      "type": "object",
      "description": "交易信息",
      "properties": {
        "delegator_address": {
          "type": "string"
        },
        "validator_address": {
          "type": "string"
        },
        "amount": {
          "type": "object",
          "properties": {
            "denom": {
              "type": "string"
            },
            "amount": {
              "type": "string"
            }
          }
        }
      }
    },
    "Tx": {
      "type": "object",
      "description": "交易信息",
      "properties": {
        "chinId": {
          "type": "string",
          "description": "chain id"
        },
        "height": {
          "type": "integer",
          "description": "高度"
        },
        "index": {
          "type": "integer",
          "description": "交易在当前块序号"
        },
        "tx_type": {
          "type": "string",
          "description": "交易类型"
        },
        "tx_type_cn": {
          "type": "string",
          "description": "交易类型中文"
        },
        "tx_detail": {
          "type": "object",
          "properties": {
            "delegate": {
              "type": "array",
              "description": "cosmos delegate",
              "items": {
                "$ref": "#/definitions/TxCosmosDelegate"
              }
            },
            "send": {
              "type": "array",
              "description": "cosmos send",
              "items": {
                "$ref": "#/definitions/TxCosmosSend"
              }
            },
            "withdraw_delegator_reward": {
              "type": "array",
              "description": "cosmos withdraw_delegator_reward",
              "items": {
                "$ref": "#/definitions/TxCosmosWithdrawDelegatorReward"
              }
            },
            "begin_unbonding": {
              "type": "array",
              "description": "cosmos begin_unbonding",
              "items": {
                "$ref": "#/definitions/TxCosmosBeginUnbonding"
              }
            }
          }
        },
        "maxgas": {
          "type": "integer",
          "description": "maxgas"
        },
        "qcp_from": {
          "type": "string",
          "description": "qcp_from"
        },
        "qcp_to": {
          "type": "string",
          "description": "qcp_to"
        },
        "qcp_sequence": {
          "type": "integer",
          "description": "qcp_sequence"
        },
        "qcp_txindex": {
          "type": "integer",
          "description": "qcp_txindex"
        },
        "qcp_isresult": {
          "type": "boolean",
          "description": "qcp_isresult"
        },
        "data": {
          "type": "string",
          "description": "原始tx数据"
        },
        "tx_status": {
          "type": "integer",
          "description": "状态"
        },
        "status": {
          "type": "integer",
          "description": "状态(0:确认中，1:成功，2:失败)"
        },
        "log": {
          "type": "string",
          "description": "日志信息"
        },
        "time": {
          "type": "string",
          "description": "UTC格式"
        }
      }
    },
...
    "Validator": {
      "type": "object",
      "description": "validator",
      "properties": {
        "name": {
          "type": "string",
          "description": "名称"
        },
        "website": {
          "type": "string",
          "description": "主页"
        },
        "identity": {
          "type": "string",
          "description": "身份"
        },
        "details": {
          "type": "string",
          "description": "备注"
        },
        "owner": {
          "type": "string",
          "description": "owner"
        },
        "chain_id": {
          "type": "string",
          "description": "chain_id"
        },
        "address": {
          "type": "string",
          "description": "address"
        },
        "pub_key_type": {
          "type": "string",
          "description": "pub_key_type"
        },
        "pub_key_value": {
          "type": "string",
          "description": "pub_key_value"
        },
        "voting_power": {
          "type": "integer",
          "description": "voting_power"
        },
        "accum": {
          "type": "integer",
          "description": "accum"
        },
        "first_block_height": {
          "type": "integer",
          "description": "first_block_height"
        },
        "first_block_time": {
          "type": "string",
          "description": "UTC"
        },
        "status": {
          "type": "integer",
          "description": "status"
        },
        "statusStr": {
          "type": "string",
          "description": "statusStr"
        },
        "inactiveCode": {
          "type": "integer",
          "description": "inactiveCode"
        },
        "inactiveTime": {
          "type": "string",
          "description": "UTC"
        },
        "inactiveHeight": {
          "type": "integer",
          "description": "inactiveHeight"
        },
        "bondHeight": {
          "type": "integer",
          "description": "bondHeight"
        },
        "percent": {
          "type": "string",
          "description": "percent"
        },
        "uptime": {
          "type": "string",
          "description": "正常运行时比"
        }
      }
    },
...
    "Status": {
      "type": "object",
      "description": "币",
      "properties": {
        "total_validators": {
          "type": "integer",
          "description": "validators节点数量"
        },
        "height": {
          "type": "integer",
          "description": "链高度"
        },
        "total_txs": {
          "type": "integer",
          "description": "交易数量"
        },
        "blockTimeAvg": {
          "type": "string",
          "description": "最近100块平均打快时间"
        }
      }
    }
  },
...
}